+++ src/Api/SoftGarden.php
@@ -5,6 +5,7 @@ namespace SWE\SoftGardenApi\Api;
 
 use GuzzleHttp\Exception\GuzzleException;
 use SWE\SoftGardenApi\ApplicantData;
+use SWE\SoftGardenApi\ApplicationData;
 use SWE\SoftGardenApi\Channel;
 use SWE\SoftGardenApi\Collection;
 use SWE\SoftGardenApi\Job;
@@ -96,13 +97,13 @@ class SoftGarden extends SoftGardenBasic
             'locale' => 'DE',
         ];
 
-        return $this->getResponse(false, $queryArguments);
+        return $this->sendResponse(self::METHOD_GET, $queryArguments);
     }
 
     /**
      * Get all channels.
      *
-     * @return Collection A collection with all channels.
+     * @return Collection<Channel> A collection with all channels.
      * @throws GuzzleException
      */
     public function getChannels(): Collection
@@ -114,7 +115,7 @@ class SoftGarden extends SoftGardenBasic
         $this->uri = 'frontend/jobslist/channels';
         $this->version = 3;
 
-        return new Collection($this->getResponse(), Channel::class, $this->useAutomaticCatalogueCompletion);
+        return new Collection($this->sendResponse(), Channel::class, $this->useAutomaticCatalogueCompletion);
     }
 
     /**
@@ -136,7 +137,7 @@ class SoftGarden extends SoftGardenBasic
         $this->uri = sprintf('frontend/jobslist/%s/job/%d', $channelId, $jobId);
         $this->version = 3;
 
-        return new Job($this->getResponse(), '', $this->useAutomaticCatalogueCompletion);
+        return new Job($this->sendResponse(), '', $this->useAutomaticCatalogueCompletion);
     }
 
     /**
@@ -161,7 +162,7 @@ class SoftGarden extends SoftGardenBasic
         ];
 
         return new JobSearchResult(
-            $this->getResponse(false, $queryArguments), '', $this->useAutomaticCatalogueCompletion
+            $this->sendResponse(self::METHOD_GET, $queryArguments), '', $this->useAutomaticCatalogueCompletion
         );
     }
 
@@ -169,7 +170,7 @@ class SoftGarden extends SoftGardenBasic
      * Get all job questions of a job.
      *
      * @param int $jobId The job id.
-     * @return Collection A collection with all questions of the job.
+     * @return Collection<JobQuestion> A collection with all questions of the job.
      * @throws GuzzleException
      */
     public function getJobQuestions(int $jobId): Collection
@@ -182,14 +183,14 @@ class SoftGarden extends SoftGardenBasic
         $this->uri = sprintf('frontend/jobs/%d/questions', $jobId);
         $this->version = 3;
 
-        return new Collection($this->getResponse(), JobQuestion::class, $this->useAutomaticCatalogueCompletion);
+        return new Collection($this->sendResponse(), JobQuestion::class, $this->useAutomaticCatalogueCompletion);
     }
 
     /**
      * Get all jobs of a channel.
      *
      * @param string $channelId The channel id.
-     * @return Collection A collection with all jobs of the channel.
+     * @return Collection<Job> A collection with all jobs of the channel.
      * @throws GuzzleException
      */
     public function getJobs(string $channelId): Collection
@@ -201,7 +202,7 @@ class SoftGarden extends SoftGardenBasic
 
         $this->uri = sprintf('frontend/jobslist/%s', $channelId);
         $this->version = 3;
-        $response = $this->getResponse();
+        $response = $this->sendResponse();
 
         return new Collection($response['results'], Job::class, $this->useAutomaticCatalogueCompletion);
     }
@@ -249,7 +250,7 @@ class SoftGarden extends SoftGardenBasic
         }
 
         return new JobSearchResult(
-            $this->getResponse(false, $queryArguments), '', $this->useAutomaticCatalogueCompletion
+            $this->sendResponse(self::METHOD_GET, $queryArguments), '', $this->useAutomaticCatalogueCompletion
         );
     }
 
@@ -266,13 +267,14 @@ class SoftGarden extends SoftGardenBasic
 
         $this->uri = 'frontend/applicants';
         $this->version = 3;
-        try {
-            $this->getResponse(true, $data);
-        } catch (GuzzleException $e) {
-            echo __LINE__;
-            echo $e->getMessage();
-            exit();
+
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Applicant Array: ' . var_export($data, true));
+            var_dump('Applicant Object: ' . var_export($applicant, true));
         }
+
+        $this->sendResponse(self::METHOD_POST, $data);
 
         return $applicant;
     }
@@ -294,14 +296,30 @@ class SoftGarden extends SoftGardenBasic
             'username' => $applicant->getUsername(),
         ];
 
-        $response = $this->getResponse(false, $fields);
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Check for username');
+            var_dump('Username: ' . $fields['username']);
+        }
+
+        $response = $this->sendResponse(self::METHOD_GET, $fields);
 
         if ($response[0] === true) {
             return true;
         }
 
         $this->uri = 'frontend/checkMailForExistence';
-        $response = $this->getResponse(false, $fields);
+        $fields = [
+            'email' => $applicant->getEmail(),
+        ];
+
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Check for email');
+            var_dump('Email: ' . $fields['email']);
+        }
+
+        $response = $this->sendResponse(self::METHOD_GET, $fields);
 
         return $response[0] === true;
     }
@@ -323,7 +341,14 @@ class SoftGarden extends SoftGardenBasic
             'password' => $applicant->getPassword(),
         ];
 
-        $response = $this->getResponse(true, $data, '', false);
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Grant Type: ' . $data['grant_type']);
+            var_dump('Username: ' . $data['username']);
+            var_dump('Password empty: ' . (empty($data['password'] ? 'true' : 'false')));
+        }
+
+        $response = $this->sendResponse(self::METHOD_POST, $data, '', false);
 
         return $response['access_token'];
     }
@@ -341,9 +366,69 @@ class SoftGarden extends SoftGardenBasic
         $this->uri = sprintf('frontend/jobs/%s/applied', $jobId);
         $this->version = 3;
 
-        $response = $this->getResponse(false, [], $uat);
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Job ID: ' . $jobId);
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        $response = $this->sendResponse(self::METHOD_GET, [], $uat);
 
         return $response[0];
+    }
+
+    /**
+     * Get all applications of an applicant.
+     *
+     * @param string $uat The user access token.
+     * @param array $queryParameters OPTIONAL. The query parameters.
+     * @return Collection<ApplicationData> Returns a collection of ApplicationData instances.
+     * @throws GuzzleException
+     */
+    public function getAllApplications(string $uat, array $queryParameters = []): Collection
+    {
+        $this->uri = 'frontend/applications';
+        $this->version = 3;
+
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            foreach ($queryParameters as $key => $value) {
+                var_dump($key . ': ' . $value);
+            }
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        return new Collection(
+            $this->sendResponse(self::METHOD_GET, $queryParameters, $uat),
+            ApplicationData::class,
+            $this->useAutomaticCatalogueCompletion
+        );
+    }
+
+    /**
+     * Get an application of an applicant.
+     *
+     * @param string $applicationId The application id.
+     * @param string $uat The user access token.
+     * @return ApplicationData Returns an ApplicationData instance.
+     * @throws GuzzleException
+     */
+    public function getApplication(string $applicationId, string $uat): ApplicationData
+    {
+        $this->uri = sprintf('frontend/applications/%s', $applicationId);
+        $this->version = 3;
+
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Application ID: ' . $applicationId);
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        return new ApplicationData(
+            $this->sendResponse(self::METHOD_GET, [], $uat),
+            '',
+            $this->useAutomaticCatalogueCompletion
+        );
     }
 
     /**
@@ -359,7 +444,13 @@ class SoftGarden extends SoftGardenBasic
         $this->uri = sprintf('frontend/applications?jobId=%s', $jobId);
         $this->version = 3;
 
-        $response = $this->getResponse(true, [], $uat);
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Job ID: ' . $jobId);
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        $response = $this->sendResponse(self::METHOD_POST, [], $uat);
 
         return $response[0];
     }
@@ -378,7 +469,13 @@ class SoftGarden extends SoftGardenBasic
         $this->uri = sprintf('frontend/applications/%s', $applicationId);
         $this->version = 3;
 
-        $this->getResponse(true, $applicationData, $uat);
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Application ID: ' . $applicationId);
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        $this->sendResponse(self::METHOD_POST, $applicationData, $uat);
     }
 
     /**
@@ -386,15 +483,68 @@ class SoftGarden extends SoftGardenBasic
      *
      * @param string $applicationId The application id.
      * @param string $uat The user access token of the applicant.
-     * @param ApplicantData $applicant The applicant instance.
+     * @param array $applicationData The application information.
      * @return void
      * @throws GuzzleException
      */
-    public function finalizeApplication(string $applicationId, string $uat, ApplicantData $applicant): void
+    public function finalizeApplication(string $applicationId, string $uat, array $applicationData = []): void
     {
         $this->uri = sprintf('frontend/applications/%s/submit', $applicationId);
         $this->version = 3;
 
-        $this->getResponse('true', $applicant->toArray(), $uat);
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Application ID: ' . $applicationId);
+            foreach ($applicationData as $key => $value) {
+                var_dump($key . ': ' . $value);
+            }
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        $this->sendResponse(self::METHOD_POST, $applicationData, $uat);
+    }
+
+    /**
+     * Delete an application if it's not finalized.
+     *
+     * @param string $applicationId The application id.
+     * @param string $uat The user access token.
+     * @return void
+     * @throws GuzzleException
+     */
+    public function deleteApplication(string $applicationId, string $uat): void
+    {
+        $this->uri = sprintf('frontend/applications/%s', $applicationId);
+        $this->version = 3;
+
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Application ID: ' . $applicationId);
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        $this->sendResponse(self::METHOD_DELETE, [], $uat);
+    }
+
+    /**
+     * Withdraw a finalized (submitted) application.
+     *
+     * @param string $applicationId The application id.
+     * @param string $uat The user access token.
+     * @return void
+     * @throws GuzzleException
+     */
+    public function withdrawApplication(string $applicationId, string $uat): void
+    {
+        $this->uri = sprintf('frontend/applications/%s/withdraw', $applicationId);
+        $this->version = 3;
+
+        if (DEBUG) {
+            var_dump(__METHOD__);
+            var_dump('Application ID: ' . $applicationId);
+            var_dump('User Access Token empty: ' . (empty($uat ? 'true' : 'false')));
+        }
+
+        $this->sendResponse(self::METHOD_POST, [], $uat);
     }
 }